pyproject.toml
[project]
name = "x-reader"
version = "0.1.0"
description = "Universal content reader — fetch, normalize, and digest content from 7+ platforms"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Leo", email = "runes.leo@gmail.com"}
]
keywords = ["content-reader", "rss", "telegram", "bilibili", "xiaohongshu", "digest"]

dependencies = [
    "requests>=2.28",
    "feedparser>=6.0",
    "python-dotenv>=1.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
telegram = ["telethon>=1.34"]
all = ["telethon>=1.34"]

[project.scripts]
x-reader = "x_reader.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["x_reader"]
